+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - This caused playout build errors like "Unable to locate history for playout item"
 - Fix crashes due to invalid smart collection searches, e.g. `smart_collection:"this collection does not exist"`
 - Fix UI crash when editing block playout that has default deco
+- Fix playback failure when seeking content with certain DTS audio (e.g. DTS-HD MA)
 
 ## [25.8.0] - 2025-10-26
 ### Added